scalafx書き換え完了
